requirements.txt
fastapi
uvicorn
python-dotenv
requests
rq
redis
openai
supabase
httpx
